// src/app/shared/generated/model/aIProvider.ts
/**
 * onecx-ai-ui-bff
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface AIProvider { 
    modificationCount?: number;
    id: string;
    name?: string;
    description?: string;
    llmUrl?: string;
    appId?: string;
    modelName?: string;
    modelVersion?: string;
    apiKey?: string;
}